Update Makefile, separate webshell
