+++ utility.h
@@ -1,0 +1,8 @@
+#ifndef UTILITY_H
+#define UTILITY_H
+
+char *escape_string_for_json(char *str);
+
+int l_strcpy(char *dest, char *src, int start, int end);
+
+#endif // UTILITY_H
